Support for multiple tags on game objects, updated object queries to leverage this for performance increases. Added worldsound class for ingame distance-based sounds

+++ Source/Scenes/Cipher/Components/CipherLock.cpp
@@ -19,8 +19,8 @@
 
 using namespace cocos2d;
 
-const int CipherLock::PinSpacing = 48.0f;
-const int CipherLock::PinUnlockDistance = 104.0f;
+const float CipherLock::PinSpacing = 48.0f;
+const float CipherLock::PinUnlockDistance = 104.0f;
 
 CipherLock* CipherLock::create()
 {
